Move workflows to config module
ErtConfig dynamically imports ErtPluginManager
to work around the plugin manager depending on
workflow-related stuff from from ert.config
