+++ src/ert/config/__init__.py
@@ -9,9 +9,12 @@ from .capture_validation import capture_validation
 from .design_matrix import DesignMatrix
 from .ensemble_config import EnsembleConfig
 from .ert_config import ErtConfig
+from .ert_plugin import ErtPlugin
+from .ert_script import ErtScript
 from .everest_constraints_config import EverestConstraintsConfig
 from .everest_objective_config import EverestObjectivesConfig
 from .ext_param_config import ExtParamConfig
+from .external_ert_script import ExternalErtScript
 from .field import Field, field_transform
 from .forward_model_step import (
     ForwardModelStep,
@@ -35,17 +38,38 @@ from .parsing import (
     WarningInfo,
 )
 from .parsing.observations_parser import ObservationType
-from .queue_config import QueueConfig
+from .queue_config import (
+    KnownQueueOptions,
+    LocalQueueOptions,
+    QueueConfig,
+)
 from .response_config import InvalidResponseFile, ResponseConfig, ResponseMetadata
 from .summary_config import SummaryConfig
 from .surface_config import SurfaceConfig
 from .workflow import Workflow
-from .workflow_job import ErtScriptWorkflow, ExecutableWorkflow, WorkflowJob
+from .workflow_config import LegacyWorkflowConfigs, WorkflowConfigs
+from .workflow_fixtures import (
+    HookedWorkflowFixtures,
+    PostExperimentFixtures,
+    PostSimulationFixtures,
+    PostUpdateFixtures,
+    PreExperimentFixtures,
+    PreFirstUpdateFixtures,
+    PreSimulationFixtures,
+    PreUpdateFixtures,
+    WorkflowFixtures,
+    fixtures_per_hook,
+)
+from .workflow_job import (
+    ErtScriptWorkflow,
+    ExecutableWorkflow,
+    WorkflowJob,
+    workflow_job_from_file,
+)
 
 __all__ = [
     "AnalysisConfig",
     "AnalysisModule",
-    "ConfigValidationError",
     "ConfigValidationError",
     "ConfigWarning",
     "DesignMatrix",
@@ -53,11 +77,14 @@ __all__ = [
     "EnsembleConfig",
     "ErrorInfo",
     "ErtConfig",
+    "ErtPlugin",
+    "ErtScript",
     "ErtScriptWorkflow",
     "EverestConstraintsConfig",
     "EverestObjectivesConfig",
     "ExecutableWorkflow",
     "ExtParamConfig",
+    "ExternalErtScript",
     "Field",
     "ForwardModelStep",
     "ForwardModelStepDocumentation",
@@ -68,8 +95,12 @@ __all__ = [
     "GenDataConfig",
     "GenKwConfig",
     "HookRuntime",
+    "HookedWorkflowFixtures",
     "InvalidResponseFile",
     "InversionTypeES",
+    "KnownQueueOptions",
+    "LegacyWorkflowConfigs",
+    "LocalQueueOptions",
     "ModelConfig",
     "ObservationGroups",
     "ObservationSettings",
@@ -77,6 +108,13 @@ __all__ = [
     "OutlierSettings",
     "ParameterConfig",
     "ParameterMetadata",
+    "PostExperimentFixtures",
+    "PostSimulationFixtures",
+    "PostUpdateFixtures",
+    "PreExperimentFixtures",
+    "PreFirstUpdateFixtures",
+    "PreSimulationFixtures",
+    "PreUpdateFixtures",
     "PriorDict",
     "QueueConfig",
     "QueueSystem",
@@ -87,8 +125,12 @@ __all__ = [
     "TransformFunction",
     "WarningInfo",
     "Workflow",
+    "WorkflowConfigs",
+    "WorkflowFixtures",
     "WorkflowJob",
     "capture_validation",
     "field_transform",
+    "fixtures_per_hook",
     "lint_file",
+    "workflow_job_from_file",
 ]
