+++ src/ert/config/external_ert_script.py
