+++ src/ert/plugins/__init__.py
@@ -2,9 +2,8 @@ from collections.abc import Callable
 from functools import wraps
 from typing import Any, ParamSpec
 
-from .ert_plugin import CancelPluginException, ErtPlugin
-from .ert_script import ErtScript
-from .external_ert_script import ExternalErtScript
+from ert.config.ert_plugin import CancelPluginException, ErtPlugin
+
 from .plugin_manager import (
     ErtPluginContext,
     ErtPluginManager,
@@ -12,20 +11,6 @@ from .plugin_manager import (
     hook_implementation,
 )
 from .plugin_response import PluginMetadata, PluginResponse
-from .workflow_config import ErtScriptWorkflow, WorkflowConfigs
-from .workflow_fixtures import (
-    HookedWorkflowFixtures,
-    PostExperimentFixtures,
-    PostSimulationFixtures,
-    PostUpdateFixtures,
-    PreExperimentFixtures,
-    PreFirstUpdateFixtures,
-    PreSimulationFixtures,
-    PreUpdateFixtures,
-    WorkflowFixtures,
-    all_hooked_workflow_fixtures,
-    fixtures_per_hook,
-)
 
 P = ParamSpec("P")
 
@@ -68,21 +53,6 @@ __all__ = [
     "ErtPlugin",
     "ErtPluginContext",
     "ErtPluginManager",
-    "ErtScript",
-    "ErtScriptWorkflow",
-    "ExternalErtScript",
-    "HookedWorkflowFixtures",
     "JobDoc",
-    "PostExperimentFixtures",
-    "PostSimulationFixtures",
-    "PostUpdateFixtures",
-    "PreExperimentFixtures",
-    "PreFirstUpdateFixtures",
-    "PreSimulationFixtures",
-    "PreUpdateFixtures",
-    "WorkflowConfigs",
-    "WorkflowFixtures",
-    "all_hooked_workflow_fixtures",
-    "fixtures_per_hook",
     "plugin",
 ]
