+++ src/ert/plugins/hook_implementations/workflows/__init__.py
@@ -12,7 +12,7 @@ from .gen_data_rft_export import GenDataRFTCSVExportJob
 from .misfit_preprocessor import MisfitPreprocessor
 
 if TYPE_CHECKING:
-    from ert.plugins.workflow_config import WorkflowConfigs
+    from ert.config.workflow_config import WorkflowConfigs
 
 
 @ert.plugin(name="ert")
