+++ tests/ert/unit_tests/plugins/test_plugin_manager.py
@@ -323,19 +323,15 @@ def test_multiple_activate_script_hook():
 
 
 def test_activate_script_plugin_integration():
-    patched = partial(
-        ert.config.ert_config.ErtPluginManager, plugins=[ActivatePlugin()]
-    )
-    with patch("ert.config.ert_config.ErtPluginManager", patched):
+    patched = partial(ert.plugins.ErtPluginManager, plugins=[ActivatePlugin()])
+    with patch("ert.plugins.ErtPluginManager", patched):
         config = ErtConfig.with_plugins().from_file_contents("NUM_REALIZATIONS 1\n")
         assert config.queue_config.queue_options.activate_script == "source something"
 
 
 def test_activate_script_plugin_integration_from_dict():
-    patched = partial(
-        ert.config.ert_config.ErtPluginManager, plugins=[ActivatePlugin()]
-    )
-    with patch("ert.config.ert_config.ErtPluginManager", patched):
+    patched = partial(ert.plugins.ErtPluginManager, plugins=[ActivatePlugin()])
+    with patch("ert.plugins.ErtPluginManager", patched):
         config = ErtConfig.with_plugins().from_dict(
             {
                 "NUM_REALIZATIONS": 1,
