+++ tests/everest/conftest.py
@@ -287,14 +287,14 @@ def no_plugins():
     patched_context = partial(
         everest.simulator.everest_to_ert.ErtPluginContext, plugins=[]
     )
-    patched = partial(ert.config.ert_config.ErtPluginManager, plugins=[])
+    patched = partial(ert.plugins.ErtPluginManager, plugins=[])
     patched_everest = partial(
         everest.config.everest_config.ErtPluginManager, plugins=[]
     )
 
     with (
         patch("everest.simulator.everest_to_ert.ErtPluginContext", patched_context),
-        patch("ert.config.ert_config.ErtPluginManager", patched),
+        patch("ert.plugins.ErtPluginManager", patched),
         patch("everest.config.everest_config.ErtPluginManager", patched_everest),
     ):
         yield
